ATOM-SG Pilot/05-Backend/requirements.txt
# ATOM-SG Pilot MVP Backend - Python Dependencies

# FastAPI framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# CORS support
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.0

# OCR processing
pytesseract==0.3.10
Pillow==10.1.0
PyMuPDF==1.23.8

# Diagram rendering
matplotlib==3.8.2

# PDF generation
reportlab==4.0.7

# Utilities
python-dateutil==2.8.2

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
